+++ src/main/java/org/apache/ibatis/reflection/wrapper/BaseWrapper.java
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2009-2021 the original author or authors.
+ *    Copyright 2009-2022 the original author or authors.
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -15,12 +15,12 @@
  */
 package org.apache.ibatis.reflection.wrapper;
 
-import java.util.List;
-import java.util.Map;
-
 import org.apache.ibatis.reflection.MetaObject;
 import org.apache.ibatis.reflection.ReflectionException;
 import org.apache.ibatis.reflection.property.PropertyTokenizer;
+
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author Clinton Begin
@@ -104,4 +104,18 @@ public abstract class BaseWrapper implements ObjectWrapper {
     }
   }
 
+  @Override
+  public boolean isCollection() {
+    return false;
+  }
+
+  @Override
+  public void add(Object element) {
+    throw new UnsupportedOperationException();
+  }
+
+  @Override
+  public <E> void addAll(List<E> list) {
+    throw new UnsupportedOperationException();
+  }
 }
